+++ tests/Frontend/e2e/Banner.cy.js
@@ -9,7 +9,7 @@ describe("Banner", function () {
       config.data.banner.title = "Banner title";
       config.data.banner.icon = "fa-solid fa-door-open";
       config.data.banner.message = "Banner text message";
-      config.data.banner.link = `${Cypress.env("redirectBaseUrl")}/?foo=a&bar=b`;
+      config.data.banner.link = `${Cypress.expose("redirectBaseUrl")}/?foo=a&bar=b`;
       config.data.banner.link_text = "Example link";
       config.data.banner.link_style = "link";
       config.data.banner.link_target = "blank";
@@ -42,7 +42,7 @@ describe("Banner", function () {
         .and(
           "have.attr",
           "href",
-          `${Cypress.env("redirectBaseUrl")}/?foo=a&bar=b`,
+          `${Cypress.expose("redirectBaseUrl")}/?foo=a&bar=b`,
         )
         .and("have.attr", "target", "_blank")
         .and("have.text", "Example link");