+++ tests/Frontend/e2e/GeneralApplication.cy.js
@@ -203,7 +203,7 @@ describe("General", function () {
 
   it("check help button if help url specified", function () {
     cy.fixture("config.json").then((config) => {
-      config.data.general.help_url = `${Cypress.env("redirectBaseUrl")}/help?foo=a&bar=b`;
+      config.data.general.help_url = `${Cypress.expose("redirectBaseUrl")}/help?foo=a&bar=b`;
 
       cy.intercept("GET", "/api/v1/config", config).as("configRequest");
     });
@@ -214,7 +214,7 @@ describe("General", function () {
       .should(
         "have.attr",
         "href",
-        `${Cypress.env("redirectBaseUrl")}/help?foo=a&bar=b`,
+        `${Cypress.expose("redirectBaseUrl")}/help?foo=a&bar=b`,
       )
       .and("have.attr", "target", "_blank")
       .invoke("removeAttr", "target");
@@ -222,7 +222,10 @@ describe("General", function () {
     cy.get('[data-test="navbar-help"]').click();
 
     // Check that redirect worked
-    cy.url().should("eq", `${Cypress.env("redirectBaseUrl")}/help?foo=a&bar=b`);
+    cy.url().should(
+      "eq",
+      `${Cypress.expose("redirectBaseUrl")}/help?foo=a&bar=b`,
+    );
   });
 
   it("check help button hidden if help url not specified", function () {