+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v4.11.0] - 2026-02-04
+
 ### Added
 
 - Configurable hashing algorithm for BBB API signatures ([#2765], [#2766])
-- `X-Robots-Tag: noindex` for all routes, excluding the landing page ([#2770])
+- `X-Robots-Tag: noindex` header for all routes, excluding the landing page ([#2770], [#2789])
+- `X-Robots-Tag: nofollow` header for all routes ([#2772], [#2789])
 
 ### Changed
 
@@ -18,10 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump redis version in docker compose files to redis 8 ([#2767])
 - Docs: Bumped the recommended PostgreSQL version to v18 ([#2769])
 
+### Removed
+
+- robots.txt file ([#2789])
+
 ### Fixed
 
 - Icon alignment inside room files tab ([#2660], [#2728])
 - Race condition during room start ([#2742])
+- Remove unnecessary `Content-Type` header from GET requests to the BigBlueButton API ([#2774], [#2775])
 
 ## [v4.10.0] - 2026-01-12
 
@@ -656,7 +664,11 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#2767]: https://github.com/THM-Health/PILOS/pull/2767
 [#2769]: https://github.com/THM-Health/PILOS/pull/2769
 [#2770]: https://github.com/THM-Health/PILOS/pull/2770
-[unreleased]: https://github.com/THM-Health/PILOS/compare/v4.10.0...develop
+[#2772]: https://github.com/THM-Health/PILOS/pull/2772
+[#2774]: https://github.com/THM-Health/PILOS/issues/2774
+[#2775]: https://github.com/THM-Health/PILOS/pull/2775
+[#2789]: https://github.com/THM-Health/PILOS/pull/2789
+[unreleased]: https://github.com/THM-Health/PILOS/compare/v4.11.0...develop
 [v3.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.0
 [v3.0.1]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.1
 [v3.0.2]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.2
@@ -677,3 +689,4 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [v4.8.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.8.0
 [v4.9.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.9.0
 [v4.10.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.10.0
+[v4.11.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.11.0